correção integer para long

+++ src/main/java/com/project/hotel_api_p2/repository/QuartoRepository.java
@@ -3,5 +3,5 @@ package com.project.hotel_api_p2.repository;
 import com.project.hotel_api_p2.entity.Quarto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuartoRepository extends JpaRepository<Quarto, Integer> {
+public interface QuartoRepository extends JpaRepository<Quarto, Long> {
 }
